pyproject.toml
[project]
name = "nexus-agent"
version = "0.3.0"
description = "Autonomous AI Quant Research & Risk Analysis Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langgraph>=0.1.0",
    "langchain-groq>=1.1.2",
    "playwright>=1.44.0",
    "beautifulsoup4>=4.12.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "speechrecognition>=3.10.0",
    "pydub>=0.25.1",
    "yfinance>=0.2.40",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
